+++ src/ai_writer/ui/components/__init__.py
@@ -2,5 +2,6 @@
 
 from ai_writer.ui.components.prompt_manager import PromptManagerDialog
 from ai_writer.ui.components.prompt_selector import PromptSelector
+from ai_writer.ui.components.settings_dialog import SettingsDialog
 
-__all__ = ["PromptManagerDialog", "PromptSelector"]
+__all__ = ["PromptManagerDialog", "PromptSelector", "SettingsDialog"]